requirements.txt
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
PyMuPDF==1.26.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-dotenv==1.1.0
